refactor: rename polygonzk, goerli and scroll files
