+++ packages/react-app-revamp/config/wagmi/index.ts
@@ -28,12 +28,12 @@ import { nearAuroraTestnet } from "./custom-chains/nearAuroraTestnet";
 import { optimism } from "./custom-chains/optimism";
 import { optimismTestnet } from "./custom-chains/optimismTestnet";
 import { polygon } from "./custom-chains/polygon";
-import { polygonTestnet } from "./custom-chains/polygonMumbai";
-import { polygonZk } from "./custom-chains/polygonZkMainnet";
+import { polygonTestnet } from "./custom-chains/polygonTestnet";
+import { polygonZk } from "./custom-chains/polygonZk";
 import { polygonZkTestnet } from "./custom-chains/polygonZkTestnet";
 import { publicGoodsNetwork } from "./custom-chains/publicGoodsNetwork";
 import { publicGoodsNetworkTestnet } from "./custom-chains/publicGoodsNetworkTestnet";
-import { scrollTestnet } from "./custom-chains/scrollGoerli";
+import { scrollTestnet } from "./custom-chains/scrollTestnet";
 import { sepolia } from "./custom-chains/sepolia";
 import { zetaTestnet } from "./custom-chains/zetaTestnet";
 import { zora } from "./custom-chains/zora";
